Fixes #3626 - Merge all RequestDispatcher related tests into one file
